Move AsyncTokenBucket to org.apache.pulsar.broker.qos package in pulsar-broker
