+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/qos/package-info.java
@@ -16,22 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.broker.service;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-import org.apache.pulsar.broker.qos.AsyncTokenBucket;
-import org.testng.annotations.Test;
-
-public class PublishRateLimiterDisableTest {
-
-    // GH issue #10603
-    @Test
-    void shouldAlwaysAllowAcquire() {
-        PublishRateLimiter publishRateLimiter = new PublishRateLimiterImpl(AsyncTokenBucket.DEFAULT_CLOCK_SOURCE);
-        Producer producer = mock(Producer.class);
-        publishRateLimiter.handlePublishThrottling(producer, Integer.MAX_VALUE, Long.MAX_VALUE);
-        verify(producer, never()).incrementThrottleCount();
-    }
-}
+/**
+ * Pulsar broker Quality of Service (QoS) related classes.
+ */
+package org.apache.pulsar.broker.qos;
