+++ pulsar-broker/src/test/java/org/apache/pulsar/broker/qos/AsyncTokenBucketTest.java
@@ -17,11 +17,12 @@
  * under the License.
  */
 
-package org.apache.pulsar.common.util;
+package org.apache.pulsar.broker.qos;
 
 import static org.testng.Assert.assertEquals;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
+import org.apache.pulsar.broker.qos.AsyncTokenBucket;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
